Improved styling
