+++ packages/esm-app-shell/src/style.css
@@ -1,0 +1,2 @@
+@import "~carbon-components/css/carbon-components.min.css";
+@import "~@openmrs/esm-styleguide/src/style.css";
