--- packages/esm-app-shell/src/style.ts
@@ -1,2 +1,0 @@
-import "carbon-components/css/carbon-components.min.css";
-import "@openmrs/esm-styleguide";
