+++ packages/esm-app-shell/src/types.ts
@@ -1,22 +1,4 @@
-declare global {
-  interface Window {
-    getOpenmrsSpaBase(): string;
-    openmrsBase: string;
-    spaBase: string;
-    spaEnv: SpaEnvironment;
-    spaVersion?: string;
-    importMapOverrides: {
-      getCurrentPageMap: () => Promise<ImportMap>;
-    };
-    installedModules: Array<any>;
-  }
-}
-
-export type SpaEnvironment = "production" | "development" | "test";
-
-export interface ImportMap {
-  imports: Record<string, string>;
-}
+import { SpaEnvironment } from "@openmrs/esm-framework";
 
 export interface SpaConfig {
   /**
