+++ packages/esm-app-shell/src/ui/breadcrumbs.tsx
@@ -3,8 +3,7 @@ import {
   Breadcrumb,
   BreadcrumbItem,
 } from "carbon-components-react/es/components/Breadcrumb";
-import { getBreadcrumbsFor } from "@openmrs/esm-api";
-import { ConfigurableLink } from "@openmrs/esm-react-utils";
+import { getBreadcrumbsFor, ConfigurableLink } from "@openmrs/esm-framework";
 
 function getPath(path: string, params: Array<string>) {
   const parts = [...params];
