+++ packages/esm-framework/src/index.ts
@@ -5,3 +5,4 @@ export * from "@openmrs/esm-extensions";
 export * from "@openmrs/esm-react-utils";
 export * from "@openmrs/esm-state";
 export * from "@openmrs/esm-styleguide";
+export * from "./types";
