+++ packages/esm-styleguide/src/index.js
@@ -1,17 +1,5 @@
-import "./canvas/canvas.css";
-import "./colors/colors.css";
-import "./typography/typography.css";
-import "./spacing/spacing.css";
-import "./cards/cards.css";
-import "./buttons-and-links/buttons-and-links.css";
-import "./logo/logo.js";
-import "./icons/icons.js";
-import "./inputs/input.css";
-import "./main-content/main-content.css";
-import "./button-icons/button-icons.css";
-import "./toasts/toasts.css";
-import "./responsive-breakpoints/responsive-breakpoints.js";
-import "./forms/forms.css";
-import "./carbon-overrides.css";
+import "./logo/logo";
+import "./icons/icons";
+import "./responsive-breakpoints/responsive-breakpoints";
 
 export * from "./toasts/toasts";
