+++ packages/esm-styleguide/src/logo/logo.js
@@ -1,4 +1,3 @@
-import "./logo.css";
 import fullLogo from "./openmrs-logo-full.svg";
 import partialLogo from "./openmrs-logo-partial.svg";
 import iconLogo from "./openmrs-logo-icon.svg";
